Move DatePicker to separate view
